+++ backend/src/main/java/com/budgetsmart/repository/UserRepository.java
@@ -8,11 +8,8 @@ import java.time.LocalDateTime;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
+public interface UserRepository extends JpaRepository<User, Integer> {
     Optional<User> findByEmail(String email);
-
     boolean existsByEmail(String email);
-
     long countByCreatedAtGreaterThanEqual(LocalDateTime date);
 }
